chore: clean up root directory and organize test files
- Remove temporary test files from root directory:
  - debug_memory.sh (debug script)
  - simple_test.sh (temporary test)
  - test_memories_read.mjs (JavaScript test)
  - test_clear (large binary file)
- Update gallery memory count type from u32 to u64
- Enable previously skipped BlobRef tests in memories_create
- Update test utilities and configurations
- Add new test files to proper tests directory structure

This cleanup ensures the root directory stays clean and all test files
are properly organized in the tests/ directory structure.

+++ src/backend/src/gallery.rs
@@ -484,7 +484,7 @@ impl Gallery {
         GalleryHeader {
             id: self.id.clone(),
             name: self.title.clone(),
-            memory_count: self.memory_entries.len() as u32,
+            memory_count: self.memory_entries.len() as u64,
             created_at: self.created_at,
             updated_at: self.updated_at,
             storage_location: self.storage_location.clone(),
